+++ data-raw/iris.R
@@ -1,0 +1,1 @@
+unbind(iris, "inst/extdata/iris")