+++ NEWS.md
@@ -1,3 +1,6 @@
+v0.0-1.1 (2014-11-07)
+===
+
 * Include an unbound version of the Iris data as example.
 * Implement `str.dfsplice` (#5).
 * During package loading, relative paths passed to splice are interpreted as relative to the package's installation directory.